requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-decouple==3.0
six==1.10.0
tokenlib==0.3.1
Werkzeug==0.12.1
WTForms==2.1
